### 2020-05-21 #2193 @whedon
[PRE REVIEW]: Hypothesize: Robust Statistics for Python
**Submitting author:** @Alcampopiano (<a href="http://orcid.org/0000-0002-3280-4447">Allan Campopiano</a>)
**Repository:** <a href="https://github.com/Alcampopiano/hypothesize" target ="_blank">https://github.com/Alcampopiano/hypothesize</a>
**Version:** v1.0.0
