### 2020-05-21 b6c0c86
added two missing DOIs
