+++ paper/paper.bib
@@ -29,14 +29,16 @@ DOI="",
   number={3},
   pages={300},
   year={1998},
-  publisher={American Psychological Association}
+  publisher={American Psychological Association},
+  DOI={ 10.1037/0003-066X.53.3.300 }
 }
 
 @book{wilcox2013introduction,
   title={Introduction to robust estimation and hypothesis testing},
   author={Wilcox, Rand R},
   year={2013},
-  publisher={Academic press}
+  publisher={Academic press},
+  DOI={ 10.1016/c2010-0-67044-1 }
 }
 
 @InProceedings{ mckinney-proc-scipy-2010,
