### 2020-05-22 #2193 @csoneson
👋@Alcampopiano - I will handle your submission and start by inviting reviewers.
Going through your paper, I think it would be useful to include a brief comparison of `Hypothesize` to other (potential) Python packages implementing similar functionality. Could you add a statement about that?
