### 2020-05-22 #2193 @csoneson
👋 @arbennett, @kirstensgithub, @martinagvilas - would you be interested in reviewing this submission for JOSS?

> Hypothesize: Robust Statistics for Python

software: https://github.com/Alcampopiano/hypothesize
accompanying paper: https://github.com/openjournals/joss-papers/blob/joss.02193/joss.02193/10.21105.joss.02193.pdf
