### 2020-05-22 #2193 @Alcampopiano
> I think it would be useful to include a brief comparison of `Hypothesize` to other (potential) Python packages implementing similar functionality. Could you add a statement about that?

@csoneson Absolutely. I will add a statement to that effect.
