### 2020-05-23 a6059f7
added statments about other useful stats libraries that include options for robust measures
