+++ paper/paper.md
@@ -30,17 +30,24 @@ bibliography: paper.bib
 
 # Summary
 
-Hypothesize is a robust statistics library for Python that is used for comparing groups and
-measuring associations. Robust methods, in particular those based on the trimmed mean [@20000755025] 
-and/or bootstrapping [@bradley1993introduction], routinely outperform traditional statistical approaches 
-in terms of power and accuracy. This is especially true when dealing with
+Hypothesize is a robust null hypothesis significance testing (NHST) library for Python.
+Robust methods, in particular those based on the trimmed mean [@20000755025] 
+and/or bootstrapping [@bradley1993introduction], routinely outperform traditional statistical 
+approaches in terms of power and accuracy. This is especially true when dealing with
 distributions that produce outliers [@wilcox1998many; @wilcox2013introduction].
 
 Hypothesize is based on Rand R. Wilcox's collection of [R functions](https://dornsife.usc.edu/labs/rwilcox/software/)
 which contains hundreds of robust methods developed since the 1960's. 
 Hypothesize brings many of these functions into the Python library ecosystem with the goal
-of making robust computations as easy as possible for researchers. 
-Hypothesize keeps the barrier-to-entry low, for example,
+of making robust hypothesis testing easy for researchers, even
+if they have not had extensive training in statistics or computer science. 
+While other statistical libraries in Python provide useful options 
+for robust model fitting [@seabold2010statsmodels] and visualizations [@ho2019moving], 
+Hypothesize focuses on providing a larger selection of robust methods 
+for comparing groups and measuring associations, especially those that are 
+analogous to traditional NHST approaches.
+
+In terms of using the software, Hypothesize keeps the barrier to entry low for researchers, for example,
 
  - To easily incorporate Hypothesize with standard data processing tools
  [see @mckinney-proc-scipy-2010], all top-level 
