+++ paper/paper.bib
@@ -41,6 +41,25 @@ DOI="",
   DOI={ 10.1016/c2010-0-67044-1 }
 }
 
+@inproceedings{seabold2010statsmodels,
+  title={statsmodels: Econometric and statistical modeling with python},
+  author={Seabold, Skipper and Perktold, Josef},
+  booktitle={9th Python in Science Conference},
+  year={2010},
+}
+
+@article{ho2019moving,
+  title={Moving beyond P values: data analysis with estimation graphics},
+  author={Ho, Joses and Tumkaya, Tayfun and Aryal, Sameer and Choi, Hyungwon and Claridge-Chang, Adam},
+  journal={Nature Methods},
+  volume={16},
+  number={7},
+  pages={565--566},
+  year={2019},
+  publisher={Nature Publishing Group},
+  DOI={ 10.1038/s41592-019-0470-3 }
+}
+
 @InProceedings{ mckinney-proc-scipy-2010,
   author    = { {W}es {M}c{K}inney },
   title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
