### 2020-05-23 #2193 @Alcampopiano
Hello @csoneson,

As you have requested, I have added the following statements (and related references) regarding other useful Python libraries that offer methods for robust statistics:

> While other statistical libraries in Python provide useful options
for robust model fitting [@seabold2010statsmodels] and visualizations [@ho2019moving],
Hypothesize focuses on providing a larger selection of robust methods
for comparing groups and measuring associations, especially those that are
analogous to traditional NHST approaches.
