### 2020-05-23 e5bfc21
added statsmodel DOI
