+++ paper/paper.bib
@@ -46,6 +46,7 @@ DOI="",
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
   year={2010},
+  DOI={ 10.25080/majora-92bf1922-011 }
 }
 
 @article{ho2019moving,
