### 2020-06-10 Alcampopiano/hypothesize#5 @martinagvilas
API documentation issues - JOSS review
This issue outlines some minor API documentation issues as part of the [review for JOSS](https://github.com/openjournals/joss-reviews/issues/2241):

It would be useful if the API documentation specified the full path to the function. E.g. instead of documenting "con1way"  as appears [here](https://alcampopiano.github.io/hypothesize/function_guide/#con1way), the full path to the function could be given: e.g. `hypothesize.utilities.con1way`. Take the API reference of [sklearn](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html#sklearn.linear_model.LogisticRegression) as an example.

This way, t would be easier for the user to import functions.
