### 2020-06-10 Alcampopiano/hypothesize#6 @martinagvilas
Software paper issues - JOSS review
This issue outlines some minor issues with the software paper as part of the [review for JOSS](https://github.com/openjournals/joss-reviews/issues/2241).

1. [x] Since the JOSS guidelines require "a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience" it would be nice to have a short definition of what a robust statistical method is before mentioning the advantages it has over other approaches.

2. [x] When mentioning the R library created by Wilcox, I believe the following citation should be added: https://link.springer.com/article/10.3758/s13428-019-01246-w

3. [x] When describing the target audience of Hypothesize, I would add that this toolbox does assume knowledge on how to perform robust analyses. The description of the methods in the API or manipulating the method's parameters assumes familiarity with concepts and acronyms used in the field (e.g. FWE, MOM). This is by no means a criticism of the toolbox per se, it is just a suggestion to better clarify who are the target users of Hypothesize.
