### 2020-06-10 b204e40
addressing issue #6 concerning JOSS paper
