+++ paper/paper.bib
@@ -71,3 +71,15 @@ DOI="",
   doi       = { 10.25080/Majora-92bf1922-00a }
 }
 
+@article{Vallat2018,
+  doi = {10.21105/joss.01026},
+  url = {https://doi.org/10.21105/joss.01026},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {31},
+  pages = {1026},
+  author = {Raphael Vallat},
+  title = {Pingouin: statistics in Python},
+  journal = {Journal of Open Source Software}
+}
