### 2020-06-10 c3dab58
addressing issue #6 concerning JOSS paper
