### 2020-06-10 e3ff76e
addressing issue #6 concerning JOSS paper
