+++ paper/paper.md
@@ -43,7 +43,7 @@ Hypothesize brings many of these functions into the Python library ecosystem wit
 of making robust hypothesis testing easy for researchers, even
 if they have not had extensive training in statistics or computer science. It is, however, assumed 
 that users have a basic understanding of the concepts and terms related to robust hypothesis 
-testing (e.g., trimmmed mean and bootstrapping).
+testing (e.g., trimmed mean and bootstrapping).
 While other statistical libraries in Python provide some useful options 
 for robust statistics [@Vallat2018; @seabold2010statsmodels; @ho2019moving], 
 Hypothesize focuses on providing a larger selection of robust methods 
