### 2020-06-10 8efd1f0
adding full path to functions in docs -  issue #5
