+++ docs/docs/function_guide.md
@@ -29,7 +29,8 @@ That is, group tests that have a single factor.
 
 ### l2drmci
 
-`#!py l2drmci(x, y, est, *args, pairwise_drop_na=True, alpha=.05, nboot=2000, seed=False)`
+`#!py hypothesize.compare_groups_with_single_factor.l2drmci
+(x, y, est, *args, pairwise_drop_na=True, alpha=.05, nboot=2000, seed=False)`
 
 Compute a bootstrap confidence interval for a
 measure of location associated with the distribution of x-y. 
@@ -89,7 +90,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### linconb
 
-`#!py linconb(x, con, tr=.2, alpha=.05, nboot=599, seed=False)`
+`#!py hypothesize.compare_groups_with_single_factor.linconb(x, con, tr=.2, alpha=.05, nboot=599, seed=False)`
 
 Compute a 1-alpha confidence interval for a set of d linear contrasts
 involving trimmed means using the bootstrap-t bootstrap method.
@@ -154,7 +155,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### pb2gen
 
-`#!py pb2gen(x, y, est, *args, alpha=.05, nboot=2000, seed=False)`
+`#!py hypothesize.compare_groups_with_single_factor.pb2gen(x, y, est, *args, alpha=.05, nboot=2000, seed=False)`
 
 Compute a bootstrap confidence interval for the
 the difference between any two parameters corresponding to two
@@ -231,7 +232,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### tmcppb
 
-`#!py tmcppb(x, est, *args, con=None, bhop=False, alpha=.05, nboot=None, seed=False)`
+`#!py hypothesize.compare_groups_with_single_factor.tmcppb(x, est, *args, con=None, bhop=False, alpha=.05, nboot=None, seed=False)`
 
 Multiple comparisons for J independent groups using trimmed means and
 the percentile bootstrap method. Rom’s method is used to control the 
@@ -299,7 +300,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### yuenbt
 
-`#!py yuenbt(x, y, tr=.2, alpha=.05, nboot=599, seed=False)`
+`#!py hypothesize.compare_groups_with_single_factor.yuenbt(x, y, tr=.2, alpha=.05, nboot=599, seed=False)`
 
 Compute a 1-alpha confidence interval for the difference between
 the trimmed means corresponding to two independent groups.
@@ -367,7 +368,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### bootdpci
 
-`#!py bootdpci(x, est, *args, nboot=None, alpha=.05, dif=True, BA=False, SR=False)`
+`#!py hypothesize.compare_groups_with_single_factor.bootdpci(x, est, *args, nboot=None, alpha=.05, dif=True, BA=False, SR=False)`
 
 Use percentile bootstrap method, compute a .95 confidence interval 
 for the difference between a measure of location or scale 
@@ -450,7 +451,7 @@ target="_blank" class="button">Try this example yourself in Colab!</a>
 
 ### rmmcppb
 
-`#!py rmmcppb(x, est, *args,  alpha=.05, con=None, dif=True, nboot=None, BA=False,
+`#!py hypothesize.compare_groups_with_single_factor.rmmcppb(x, est, *args,  alpha=.05, con=None, dif=True, nboot=None, BA=False,
[…]
