### 2020-06-10 fc8d388
indicating positional only arguments where applicable in docs and docstrings- issue #4
