### 2020-06-11 #2241 @martinagvilas
Thanks @Alcampopiano!

@csoneson all my issues have been addressed, so I recommend the acceptance of Hypothesize.
