### 2020-06-16 Alcampopiano/hypothesize#8 @arbennett
Provide documentation and dependencies for running tests
The contributing documentation describes how tests are structured and tools used, but doesn't actually say how to run the tests. I think adding a sentence or two describing how to run the tests would be useful. Perhaps something like:

> Following installation, to run the test suite first navigate to the `tests` directory then use the `pytest` command.

Additionally, in the `setup.py` it would be useful to add `tests_require=['pytest']` for completeness.
