### 2020-06-16 0657d7d
added assert_allclose for numpy arrays and assert_frame_equal (with less precise checking) to the testing module; issue #7
