### 2020-06-16 Alcampopiano/hypothesize#7 @Alcampopiano
Hello @arbennett,

This makes so much sense. Thank you very much for pointing this out to me. I've added more tolerant assertions now as you've suggested (for arrays and DataFrames). I will close this issue but feel free to reopen if I have not addressed your suggestion sufficiently.
