### 2020-06-16 759ef49
sentence added to indicate how to run the tests using pytest; issue #8
