### 2020-06-16 Alcampopiano/hypothesize#8 @Alcampopiano
Hello @arbennett,

Thank you for this suggestion. I have added the following sentence as suggested:

> To run the test suite, first navigate to the "tests" directory then use the `pytest` command from your terminal.

I will close this issue but feel free to reopen if I have not fully addressed the issue.
