### 2020-06-16 87cbd85
added pytest as a testing requirement in setup file; issue #8
