### 2020-06-16 Alcampopiano/hypothesize#9 @arbennett
JOSS paper suggestions
Hi @Alcampopiano - thanks for your work in creating hypothesize! I think it's a really nice piece of software and a good contribution to JOSS. However I think the paper in its current form could be improved to make it more descriptive and stand on it's own a bit better. These are my suggestions:

1. In the text of the paper you list a couple of other packages for robust statistics, and say that hypothesize implements more methods. Could you elaborate on what was missing from these other packages and/or what problems hypothesize can solve that these others cannot?

2. I think your documentation is really great, and you could bring some of this information into the paper as well. Namely, I think including the package structure and list of estimators from  the overview page would be really helpful for making it clear what is in hypothesize.
