### 2020-06-16 #2241 @arbennett
Hypothesize is a well designed library with a clear and well defined scope as well as functionality that does not appear to exist in other Python packages. I think this is a good contribution and a good fit to be published in JOSS after some minor modifications.

I really enjoyed how clean the documentation and structure of the code is. I had almost no trouble installing and getting started with hypothesize. The only functional issue that I had was three minor points on testing. I have opened issues [here](https://github.com/Alcampopiano/hypothesize/issues/7) and [here](https://github.com/Alcampopiano/hypothesize/issues/8) describing these issues.

Additionally, I think that some of the text in the paper could be elaborated on and expanded to make it clearer why user should use hypothesize over some of the other referenced packages, and a bit more information about what is actually implemented in hypothesize. Most of this _is_ found in the documentation, but I think it would also be useful to include in the paper itself. I have opened an issue [here](https://github.com/Alcampopiano/hypothesize/issues/9) specifically describing my suggestions.
