### 2020-06-16 626dd1b
adding a soft contrast in the paper between Hypothesize and other related packages; Including a focused list to the paper to describe key features in Hypothesis; issue #9
