### 2020-06-16 Alcampopiano/hypothesize#9 @Alcampopiano
Hello @arbennett,

Thank you for these suggestions. I believe they have indeed lead to an improved manuscript.

1. I have adding content to the paper to address how Hypothesize is different from other stats packages, but in a very general and implicit way. I chose to simply mention that the main reason Hypothesize is special is that it is solely dedicated to robust methods (i.e., estimators and hypothesis tests). No other Python library, that I know of, has this restricted focus.

> In contrast to other statistical libraries in Python [@Vallat2018; @seabold2010statsmodels; @ho2019moving], Hypothesize is focused solely on robust methods, especially those that are analogous to traditional NHST approaches (e.g., t-test, ANOVA, Pearson's correlation).

My reasoning for not writing a larger compare-and-contrast is that each of the stats packages listed in the paper offer a unique (but partially overlapping) feature set in terms of robust statistics and believe that if I were to explain the differences fairly, there would be too much detail in the paper. For example,

- statsmodels offers several exotic estimators but has limited methods for conducting robust hypothesis tests (as far as I can see). And the API for such tests, in my opinion, is harder to use for those learning statistics or those who are not experienced programmers.

- Pingouin offers a very clear API in general (even for beginner researchers in my opinion) but little in the way of robust functionality. It does, however, allow robust FWE control (which I don't believe statsmodel offers).

- Dabest has a large focus on visualization, but a limited set of supported statistical designs as far as I can see (e.g., you cannot do an ANOVA-like test, or correlations). It does, however, offer effect sizes, which Hypothesize does not do nearly as well.

Please let me know if you still feel that I need more detail in terms of a compare-and-contrast. I appreciate your suggestions. Also note that I believe point 2 helps to describe what makes Hypothesize different without being too forceful about it.

---

2. Thank you for this wonderful suggestion. I really appreciate it as it adds a lot to the paper. I wrote:

> Hypothesize also provides many flexible options for controlling how hypothesis tests are carried out.
> For example,
>
> - Several estimators are available, including the trimmed mean,
>     mean, median, winsorized correlation and percentage bend correlation
>     (estimators from other statistics packages, or user defined estimators, can also be used in some cases).
>
> - For many designs, users can choose between using the percentile bootstrap or the bootstrap-t test.
>
> - To control family-wise error, robust sequentially rejective methods can be applied.
>

I will leave this open until you close it or give me the "okay".

Thank you again.
