### 2020-06-17 Alcampopiano/hypothesize#9 @arbennett
Thanks for the thoughtful response, I agree with you, the text that you've added with respect to my second point is a good way to clarify both points. I would be happy to accept these changes and move forward with the review!
