### 2020-06-17 86d96ce
refining the paper's content around what makes Hypothesize attractive; issue #9
