+++ paper/paper.bib
@@ -83,3 +83,34 @@ DOI="",
   title = {Pingouin: statistics in Python},
   journal = {Journal of Open Source Software}
 }
+
+@article{rom1990sequentially,
+  title={A sequentially rejective test procedure based on a modified Bonferroni inequality},
+  author={Rom, Dror M},
+  journal={Biometrika},
+  volume={77},
+  number={3},
+  pages={663--665},
+  year={1990},
+  publisher={Oxford University Press}
+}
+
+@article{hochberg1988sharper,
+  title={A sharper Bonferroni procedure for multiple tests of significance},
+  author={Hochberg, Yosef},
+  journal={Biometrika},
+  volume={75},
+  number={4},
+  pages={800--802},
+  year={1988},
+  publisher={Oxford University Press}
+}
+
+@article{benjamni1995controlling,
+  title={Controlling the false discovery rate: a practical and powerful approach to multiple testing},
+  author={Benjamni, Y and Hochberg, Y and Benjamini, Y},
+  journal={JR Stat Soc B},
+  volume={57},
+  pages={289--300},
+  year={1995}
+}
