### 2020-06-17 #2241 @csoneson
@Alcampopiano - I see that you have closed the issues raised by @arbennett above, and there was a confirmation from @arbennett in at least one of them. Could you both please confirm if you consider these issues to be satisfactorily addressed and we should go ahead with accepting the submission, or if there are still open issues? Thanks!
