+++ paper/paper.bib
@@ -108,7 +108,7 @@ DOI="",
 
 @article{benjamni1995controlling,
   title={Controlling the false discovery rate: a practical and powerful approach to multiple testing},
-  author={Benjamni, Y and Hochberg, Y and Benjamini, Y},
+  author={Benjamni, Y and Hochberg},
   journal={JR Stat Soc B},
   volume={57},
   pages={289--300},
