### 2020-06-17 34f6bee
add missing DOIs
