+++ paper/paper.bib
@@ -92,7 +92,8 @@ DOI="",
   number={3},
   pages={663--665},
   year={1990},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI={10.1093/biomet/77.3.663}
 }
 
 @article{hochberg1988sharper,
@@ -103,7 +104,8 @@ DOI="",
   number={4},
   pages={800--802},
   year={1988},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI = {10.1093/biomet/75.4.800}
 }
 
 @article{benjamni1995controlling,
@@ -112,5 +114,6 @@ DOI="",
   journal={JR Stat Soc B},
   volume={57},
   pages={289--300},
-  year={1995}
+  year={1995},
+  DOI = {https://doi.org/10.1111/j.2517-6161.1995.tb02031.x}
 }
