### 2020-06-17 #2241 @Alcampopiano
Hello @csoneson,

Are you able to check the references again when you have a chance (I don't believe I can ask Whedon to do that)? I have attempted to fix the missing DOIs.
