### 2020-06-17 03dc34a
add missing DOIs
