+++ paper/paper.bib
@@ -115,5 +115,5 @@ DOI="",
   volume={57},
   pages={289--300},
   year={1995},
-  DOI = {https://doi.org/10.1111/j.2517-6161.1995.tb02031.x}
+  DOI = {doi.org/10.1111/j.2517-6161.1995.tb02031.x}
 }
