### 2020-06-17 #2241 @Alcampopiano
@csoneson Thank you. I could not get the last one to resolve in an online DOI checker, so I added the full URL. That is invalid, which makes sense.

Okay, I've added the DOI as Whedon has suggested and I hope that will work.
