### 2020-06-17 #2241 @csoneson
@Alcampopiano - Could you make the following small fixes to the references:
* the DOI for Benjamini & Hochberg (1995) should not contain `doi.org/`
* Benjamini & Hochberg (1995) is missing an "i" in Benjamini
* There seem to be spaces in the DOIs for Ho et al (2019), McKinney (2010), Seabold & Perktold (2010), Wilcox (1998) and Wilcox (2013)
* Hochberg (1988) and Rom (1990) need case protection in the title to keep the capital B in Bonferroni
* For consistency, please write out the full journal name also for Benjamini & Hochberg (1995) and check the capitalization of journal names.
Note that if you click on the article proof above and then download the pdf, you can check that the DOI links work by clicking on them directly in the bibliography.
