### 2020-06-17 916ef44
fixing minor reference details and formatting
