+++ paper/paper.bib
@@ -16,7 +16,7 @@ DOI="",
 @article{bradley1993introduction,
   title={An introduction to the bootstrap},
   author={Bradley, Efron and Tibshirani, Robert J},
-  journal={Monographs on statistics and applied probability},
+  journal={Monographs on Statistics and Applied Probability},
   volume={57},
   year={1993}
 }
@@ -30,7 +30,7 @@ DOI="",
   pages={300},
   year={1998},
   publisher={American Psychological Association},
-  DOI={ 10.1037/0003-066X.53.3.300 }
+  DOI={10.1037/0003-066X.53.3.300}
 }
 
 @book{wilcox2013introduction,
@@ -38,7 +38,7 @@ DOI="",
   author={Wilcox, Rand R},
   year={2013},
   publisher={Academic press},
-  DOI={ 10.1016/c2010-0-67044-1 }
+  DOI={10.1016/c2010-0-67044-1}
 }
 
 @inproceedings{seabold2010statsmodels,
@@ -46,11 +46,11 @@ DOI="",
   author={Seabold, Skipper and Perktold, Josef},
   booktitle={9th Python in Science Conference},
   year={2010},
-  DOI={ 10.25080/majora-92bf1922-011 }
+  DOI={10.25080/majora-92bf1922-011}
 }
 
 @article{ho2019moving,
-  title={Moving beyond P values: data analysis with estimation graphics},
+  title={Moving beyond P values: Data analysis with estimation graphics},
   author={Ho, Joses and Tumkaya, Tayfun and Aryal, Sameer and Choi, Hyungwon and Claridge-Chang, Adam},
   journal={Nature Methods},
   volume={16},
@@ -58,17 +58,17 @@ DOI="",
   pages={565--566},
   year={2019},
   publisher={Nature Publishing Group},
-  DOI={ 10.1038/s41592-019-0470-3 }
+  DOI={10.1038/s41592-019-0470-3}
 }
 
-@InProceedings{ mckinney-proc-scipy-2010,
-  author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
-  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
-  pages     = { 56 - 61 },
-  year      = { 2010 },
-  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
-  doi       = { 10.25080/Majora-92bf1922-00a }
+@InProceedings{mckinney-proc-scipy-2010,
+  author    = {{W}es {M}c{K}inney },
+  title     = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+  booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+  pages     = {56 - 61},
+  year      = {2010 },
+  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  doi       = {10.25080/Majora-92bf1922-00a}
 }
 
 @article{Vallat2018,
@@ -80,12 +80,12 @@ DOI="",
   number = {31},
   pages = {1026},
   author = {Raphael Vallat},
-  title = {Pingouin: statistics in Python},
+  title = {Pingouin: Statistics in Python},
   journal = {Journal of Open Source Software}
 }
 
 @article{rom1990sequentially,
-  title={A sequentially rejective test procedure based on a modified Bonferroni inequality},
+  title={A sequentially rejective test procedure based on a modified {B}onferroni inequality},
   author={Rom, Dror M},
   journal={Biometrika},
   volume={77},
@@ -97,7 +97,7 @@ DOI="",
 }
 
 @article{hochberg1988sharper,
[…]
