### 2020-06-17 7ae724f
fixing minor reference details and formatting
