+++ paper/paper.bib
@@ -80,7 +80,7 @@ DOI="",
   number = {31},
   pages = {1026},
   author = {Raphael Vallat},
-  title = {Pingouin: Statistics in Python},
+  title = {Pingouin: Statistics in {P}ython},
   journal = {Journal of Open Source Software}
 }
 
