### 2020-06-17 f5afe29
remove hyphen in JOSS
