### 2020-06-17 #2241 @Alcampopiano
@csoneson Thank you for pointing out the various things to fix and check in the reference section.

I believe I have addressed each of the points listed above (and I have verified that each DOI is valid by clicking on them in the proof).
