### 2020-06-18 #2241 @csoneson
@Alcampopiano - great, thanks for making those edits! I found one more detail - in "Bradley & Tibshirani (1993)", Bradley Efron's first name and surname have been swapped - it should be "Efron & Tibshirani (1993)". Could you fix that too?

The next steps:
- [ ] Make a tagged release of the software and list the version tag of the archived version here.
- [ ] Archive the software in Zenodo.
- [ ] Check the metadata of the Zenodo archive - it should have exactly the same title and authors as the paper. You can also add authors' ORCiD there if you want.
- [ ] Report the DOI of the Zenodo archive here
