### 2020-06-18 6a94dee
version 1.1.0; reference fixes
