+++ paper/paper.bib
@@ -15,7 +15,7 @@ DOI="",
 
 @article{bradley1993introduction,
   title={An introduction to the bootstrap},
-  author={Bradley, Efron and Tibshirani, Robert J},
+  author={Efron, Bradley and Tibshirani, Robert J},
   journal={Monographs on Statistics and Applied Probability},
   volume={57},
   year={1993}
