### 2020-06-18 #2241 @Alcampopiano
Hello @csoneson,

Thank you very much for your incredibly careful eye on the paper. I have fixed the citation so that the name appears correctly.

In addition:

- I have made a tagged release: `v1.1.0`
- I have archived that release in Zenodo
- The title and authors in the paper match Zenodo's corresponding metatdata exactly
- Zenodo DOI: http://doi.org/10.5281/zenodo.3900100

Please let me know if you need anything else.
