### 2020-06-19 #2241 @Alcampopiano
@martinagvilas @arbennett @csoneson @arfon

Publishing in JOSS was an incredibly positive experience for me. My sincerest thanks to you all. I really appreciate all of your hard work. I made sure to acknowledged Whedon in the paper too :).
